UserTable Refactor
